requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
pdfplumber==0.11.0
openpyxl==3.1.4
python-multipart==0.0.9
anthropic==0.28.0
jinja2==3.1.4
scipy==1.13.1
razorpay
mysql-connector-python
python-dotenv==1.0.1
reportlab
sqlalchemy
pymysql